simple_erp/erp_core/serializers/invoice_item.py
from rest_framework import serializers
from ..models import InvoiceItem


class InvoiceItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = InvoiceItem
        fields = ['id', 'product', 'quantity']
