simple_erp/requirements.txt
Django
mysqlclient
django-mysql
djangorestframework